Initialise gcc-lua project
gcc‑lua extends the GNU Compiler Collection with the ability to run Lua
scripts. The plugin provides an interface to register callback functions
for plugin events, and inspect the abstract syntax tree of a translation
unit. The plugin is useful for static C code analysis. gcc‑lua supports
GCC 4.6 or 4.7, and, Lua 5.1 or 5.2, or LuaJIT.